requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
mysql-connector-python==8.0.24
mysqlclient==2.0.3
protobuf==3.15.8
psycopg2==2.8.6
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
